Translatable configurable metadata
This is a continuation of #1103 where we made the Organization-data in metadata
configurable using the translations-files.

However, the result would be like this:

<md:OrganizationDisplayName lang="en">My fictional organization</md:OrganizationDisplayName>
<md:OrganizationDisplayName lang="nl">My fictional organization</md:OrganizationDisplayName>

Instead of:

<md:OrganizationDisplayName lang="en">My fictional organization</md:OrganizationDisplayName>
<md:OrganizationDisplayName lang="nl">Mijn fictieve organisatie</md:OrganizationDisplayName>

Everything was being translated using the default locale. This PR attempts to fix that.

See: https://github.com/OpenConext/OpenConext-engineblock/pull/1292
See: https://github.com/OpenConext/OpenConext-engineblock/pull/1103