Locale is considered when providing keywords to WAYF
WHen constructing the WAYF data, previously the current locale was
not considered and the method would just in order use the first
available of en, nl, or pt. With this change, Engine will first
see if there are keywords for the current locale of the WAYF, if
not, use en as a fallback, otherwise return Undefined as before.