Prevent adding empty renamed attribute
Issue: #1321 describes it perfectly. When Manage configures an override
for an attribute name. But the targetted attribute is not in the assertion, then
Engine would add an empty attribute with the overridden value to the
assertion. The SAML2 lib would then warn us that an empty attribute
value can not be processed.

To fix that. I added a test statement that verifies if the targetted
attribute is present in the assertion. If not, the substitution is not
made and a warning is logged.

It is not allowed to rename an attribute with a null value.
It is allowed to rename an empty attribute.

See: https://github.com/OpenConext/OpenConext-engineblock/issues/1321