Feat: Implement Discovery support
Prior to this change, there were cases where it wasn't clear which
IdP end users should use. In these scenario's the users needed an IdP
which was not recognisable for them.

This change adds support for discovery IdP entries.
Which are additional names / ways of finding an IdP in the WAYF.
These can be configured in Manage.

A discovery requires at least an english name, but can also include
keywords or a custom logo, which is used on the consent page as well.

Resolves https://github.com/OpenConext/OpenConext-engineblock/issues/1338