Upgrade behat/mink-goutte-driver and friends-of-behat/mink-extension
Prior to this change, SF could not be upgraded to SF54, because behat/mink-goutte-driver would hold it back.

In order to upgrade that package, friends-of-behat/mink-extension also needs an update, because otherwise the fabpot/goutte driver could not be instantiated.

In order to make this possible, composer needs to think it runs on php 7.4, as it is listed as a requirement in one of those packages.