Make codebase PHP 8.2 compatible
Prior to this change, EB would not run on PHP 8.2.
This change does the minimal needed changes to make EB run on PHP 8.2.

Mainly, work around the stricter serialization / deserialization of
objecs introduced in PHP 8.1.

---

PHP55: use class constant

RandomFunctionRector
ThisCallOnStaticMethodToStaticCallRector

PHP71 RemoveExtraParametersRector

PHP72 StringsAssertNakedRector

PHP73 SensitiveHereNowDocRector

GetDebugTypeRector
OptionalParametersAfterRequiredRector

PHP82 - Fix unset warnings without bcbreak.

Upgrade pimple to php8 compatible version

composer update openconext/saml-value-object friends-of-behat/symfony-extension ingenerator/behat-tableassert doctrine/doctrine-migrations-bundle -W -m

PHP82 in pipeline

Fix: Error: Call to a member function authenticate() on null

Prior to this change, the authorization mechanism could try to trigger a re-authentication. (because it did not trust the token?)

This caused the Connections, Consent, Deprovision and AttributeReleasePolicy controller tests fail.

This change prevents re-authentication attempts in the authorization flow by performing them upfront.

Fix Loopguard serialization issue

Prior to this change, some behat tests, like `AcsTinkering.feature` would fail when redirecting to the error page.
This was caused by the FunctionalTestingAuthenticationLoopGuard being serialized into the session.
This loopguard only needs the three int values in its state. But the Func testing variant had the datastore, including flysystem, serialized into the session. This caused gateway timeouts, causing the tests to fail.

This change removes the storage from the scope of the object, and injects is when needed from the context.

Do not serialize certificate resources in the SP or IDP to the session.

In the consent flow, the IDP/SP is serialized to the session via the 'role' property.
This change enables serialization of these entities in PHP 8.1, where the serialization of a resource is no longer possible.
Technically, it was never possible, as resources cannot be serialized.

Enable serialization of Request / Response to session

Prior to this change, PHP 8.1 would not allow the AuthnRequest or the Response to be serialized to the session, because it contains a DomDocument.
This change converts the underlying message to XML on serialization, and restores the original message from XML on deserialisation to the session.

PHP81: Make child classes conform to parent signature

Update monolog/guzzle and make depencency on robrichards/xmlseclibs explicit.

Resolve multiple Doctrine Entity Managers
Prior to this change, the tests would fail.
There were mulitple services that were essentially the same entity manager. But because they were separate, calling `persist` on a service, one instance of the EM would try to insert a new instance, as it was untracked in that instance of the EM.

This change injects the same EM in all services, eliminating multiple state issues.
However, I have no idea why the extra `persist` calls are needed. But many tests fail if they are absent.
(Also, what was the reason for the compat EM in the first place?)

Package updates & fix deprecations

Package updates & prepare for doctrine upgrade by changing json_array > json

Make phpunit trigger & show deprecations

Fix deprecations triggered during phpunit tests.

Move packages back to no-dev requirements

Packages were moved by accident to --dev

Upgrade `openconext/saml-value-object` to the new 3.0 release for php 8.2 support via upgraded `beberlei/assert`.

On production, the column type is `array`, not `json_array`. This is not needed.

Restore accidentaly removed var/cache/.gitkeep

Instead of using reflection, return a static array of properties to serialize.
Because it might be a slight performance improvement.

Add routes/prod to ensure bootstrapping succeeds for --env=prod

Prevent error on accessing unset logo

Preserve RelayState through serialization
Prior to this change, the RelayState would get lost during serialization, this caused the IdP to not accept the response.
This change ensures the RelayState gets preserved during the login flow.