Fix Loopguard serialization issue
Prior to this change, some behat tests, like `AcsTinkering.feature` would fail when redirecting to the error page.
This was caused by the FunctionalTestingAuthenticationLoopGuard being serialized into the session.
This loopguard only needs the three int values in its state. But the Func testing variant had the datastore, including flysystem, serialized into the session. This caused gateway timeouts, causing the tests to fail.

This change removes the storage from the scope of the object, and injects is when needed from the context.