Resolve multiple Doctrine Entity Managers
Prior to this change, the tests would fail.
There were mulitple services that were essentially the same entity manager. But because they were separate, calling `persist` on a service, one instance of the EM would try to insert a new instance, as it was untracked in that instance of the EM.

This change injects the same EM in all services, eliminating multiple state issues.

However, I have no idea why the extra `persist` calls are needed. But many tests fail if they are absent.

(Also, what was the reason for the compat EM in the first place?)