Fix/rework building of javascript
Prior to this change, in the previous commit, the `tinyify` package was
upgraded. This package minified the js and css.
However, after the update, the javascript was no longer functioning.
This manifested itself by errors in the browser console.

This had two root causes:
 * The javascript in EB used circular dependencies.
 * Even after fixing that, the tinyify package simply broke the
   javascript by removing functions that could not be removed.

This change reworks the javascript to remove the circular dependencies.
It also replaces the unmaintained tinyify plugin by using the `terser` directly.