Ensure Stepup is performed before SRAM
Prior to this change, stepup would be performed after the SRAM interrupt check.
This change ensures steupup is performed first.
The reason is that the SRAM callout could trigger a 2fa check in SBS. It would be weird to perform a 2fa stepup after that 2fa check.
So doing the stepup first would seem better.