Fix hardcoded parameters being duplicated and not overridable
Prior to this change, the parameter overrides from various environments did not work.
Also, it was very unclear which config was used, as for example, asset_version was declared both in `config/packages/config.yaml` and in `config/services.yaml`.
The parameters in `config/packages/config.yaml` were never read.

This change removes the parameters in `config/services.yaml`, as environment specific overrides did not work if these were defined. So for `--env=dev`, it was not `dev`, but `#ASSET_VERSION#`.

This change fixes that, and also moves all static parameters to `framework.yaml` for consistency. (not to `parameters.yaml` as it is used for non-hardcoded parameters).

https://symfony.com/doc/current/configuration.html#configuration-parameters